### 2025-09-30 #9108 @editorialbot
[PRE REVIEW]: msibi: Multistate Iterative Boltzmann Inversion
**Submitting author:** @chrisjonesBSU (<a href="http://orcid.org/0000-0002-6196-5274">Chris D. Jones</a>)
**Repository:** https://github.com/mosdef-hub/msibi
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.0
