### 2025-09-30 f8465c6
fix DOIs, remove acknowledgement heading
