+++ paper/paper.bib
@@ -5,7 +5,7 @@
     year="2014",
     volume="140",
     number="22",
-    doi="http://dx.doi.org/10.1063/1.4880555"
+    doi="10.1063/1.4880555"
 }
 
 @article{Anderson2020hoomd,
@@ -28,7 +28,7 @@
   pages={786--803},
   year={2021},
   publisher={Taylor \& Francis},
-  doi={https://doi.org/10.1080/08927022.2020.1828583}
+  doi={10.1080/08927022.2020.1828583}
 }
 
 @article{Carbone2008,
@@ -39,7 +39,7 @@
   number={6},
   year={2008},
   publisher={AIP Publishing},
-  doi={https://doi.org/10.1063/1.2829409}
+  doi={10.1063/1.2829409}
 }
 
 @article{Moore2016,
@@ -51,7 +51,7 @@
   pages={9944--9958},
   year={2016},
   publisher={ACS Publications},
-  doi={https://doi.org/10.1021/acs.jpcb.6b08046}
+  doi={10.1021/acs.jpcb.6b08046}
 }
 
 @article{Jones2025,
@@ -63,7 +63,7 @@
   pages={117},
   year={2025},
   publisher={MDPI},
-  DOI = {https://doi.org/10.3390/polym17010117}
+  DOI = {10.3390/polym17010117}
 }
 
 @article{Tang2023,
@@ -75,7 +75,7 @@
   pages={2256--2264},
   year={2023},
   publisher={Royal Society of Chemistry},
-  doi ={https://doi.org/10.1039/D2CP04849C}
+  doi ={10.1039/D2CP04849C}
 }
 
 @article{Fritz2009,
@@ -87,7 +87,7 @@
   pages={7579--7588},
   year={2009},
   publisher={ACS Publications},
-  doi={https://doi.org/10.1021/ma901242h}
+  doi={10.1021/ma901242h}
 }
 
 @article{Martini2007,
@@ -99,7 +99,7 @@
   pages={7812--7824},
   year={2007},
   publisher={ACS Publications},
-  doi={https://doi.org/10.1021/jp071097f}
+  doi={10.1021/jp071097f}
 
 }
 
@@ -123,7 +123,7 @@
   pages = {109129},
   year = {2020},
   issn = {0927-0256},
-  doi = {https://doi.org/10.1016/j.commatsci.2019.109129},
+  doi = {10.1016/j.commatsci.2019.109129},
   url = {https://www.sciencedirect.com/science/article/pii/S0927025619304203},
   author = {Eric Jankowski and Neale Ellyson and Jenny W. Fothergill and Michael M. Henry and Mitchell H. Leibowitz and Evan D. Miller and Mone’t Alberts and Samantha Chesser and Jaime D. Guevara and Chris D. Jones and Mia Klopfenstein and Kendra K. Noneman and Rachel Singleton and Ramon A. Uriarte-Mendoza and Stephen Thomas and Carla E. Estridge and Matthew L. Jones},
 }
@@ -153,7 +153,7 @@
   pages={1624--1636},
   year={2003},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1002/jcc.10307}
+  doi={10.1002/jcc.10307}
 }
 
 @article{Thompson2022,
@@ -164,7 +164,7 @@
   pages={108171},
   year={2022},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.cpc.2021.108171}
+  doi={10.1016/j.cpc.2021.108171}
 }
 
 
@@ -177,7 +177,7 @@
   pages={1701--1718},
   year={2005},
   publisher={Wiley Online Library},
[…]
