### 2025-10-01 #9108 @chrisjonesBSU
There is one item I feel worth mentioning in case it is relevant for the editor and/or reviewers. The mosdef-hub msibi repository used for this submission was originally created in 2014 and contained some initial code used in the first paper that introduced the MS-IBI method itself (as opposed to the software created). After sitting dormant for several years, I began revamping msibi for my PhD project in a fork.  During this time, I became an admin and developer with [mosdef-hub](https://github.com/mosdef-hub). After discussing with everyone, we decided to push this work back upstream to the mosdef-hub msibi repository.

The pull-request, which can be seen [here](https://github.com/mosdef-hub/msibi/pull/39), essentially resulted in a complete overhaul and revamp of the msibi code-base, containing 629 commits and 156 files changed. The current authors include those who have made significant commits over the last 3-years or vital contributions through guiding development and writing of the manuscript.
