### 2025-10-01 274c59c
Fix typos in paper
