+++ paper/paper.md
@@ -36,7 +36,7 @@ bibliography: paper.bib
 Iterative Boltzmann inversion (IBI) is a well-established, and widely used, method for deriving coarse-grained (CG) force fields that recreate the structural distributions of an underlying atomistic model.
 Multiple state IBI (MS-IBI) as introduced by Moore et al. [@Moore2014], addresses state-point transferability limitations of IBI by including distributions from multiple state points to inform the derived CG force field.
 Here, we introduce `msibi`, a pure python package that implements the MS-IBI method for creating CG force fields for both intramolecular and intermolecular interactions.
-The package offers a user-friendly, Python-native API, eliminating the need for bash scripting and manual editing of multiple input files.
+The package offers a user-friendly, Python-native API, eliminating the need for bash scripting and manual editing of input files.
 `msibi` is ultimately simulation engine agnostic, but uses the HOOMD-Blue simulation engine [@Anderson2020hoomd] under-the-hood to perform iterative CG model simulations.
 This allows `msibi` to utilize graphical processing unit (GPU) acceleration without requiring users to manually compile GPU compatible code.
 
@@ -45,10 +45,10 @@ This allows `msibi` to utilize graphical processing unit (GPU) acceleration with
 Molecular dynamics (MD) simulations are computationally intensive and scale poorly with the number of particles in the system, which limits accessible time and length scales.
 As a result, atomistic MD simulations of complex systems such as polymers and biomolecules become prohibitively expensive, especially as their relevant length and time scales often surpass micrometers and microseconds.
 Coarse-graining (CG) is a commonly adopted solution to this challenge, as it reduces computational cost by grouping—or mapping—atoms into a single, larger bead [@Joshi2021].
-However, this approach introduces two challenges: first, the potential energy surface for a given chemistry and CG mapping is not known a priori, and
+However, this approach introduces two challenges: first, the potential energy surface for a given chemistry and CG mapping is not known *a priori*, and
 second, as the mapping used is arbitrary, with multiple valid options, developing a single CG force field that is transferable across various mapping choices is not possible.
 Also, a CG force field derived using IBI is state-point dependent with limited transferability to other state-points [@Moore2014; @Carbone2008].
-Consequently, developing a CG force field is required each time a new under-lying chemistry, mapping or target state-point is required.
+Consequently, developing a CG force field is required each time a new under-lying chemistry, mapping or target state-point is used.
 IBI and MS-IBI are popular choices for deriving CG forces for polymers and biomolecules [@Carbone2008; @Moore2016; @Jones2025; @Tang2023; @Fritz2009].
 While these methods are frequently used, open-source software tools that provide an accessible and reproducible, end-to-end workflow for IBI and MS-IBI remain limited, especially for arbitrary mappings and multi-state systems.
 
@@ -63,7 +63,7 @@ This approach follows best practices for deriving CG force fields via IBI and MS
 Additionally, we emphasize that `msibi` is ultimately engine agnostic: any simulation engine can be used to generate the fine-grained target structural distributions, and the CG force field produced by `msibi` is compatible with any simulation engine that supports tabulated forces.
 This includes LAMMPS [@Thompson2022], Gromacs [@Van2005], DL_Poly [@Smith2002], and HOOMD-Blue [@Anderson2020hoomd], among others.
 It is required that the target trajectories are converted to the [gsd](https://gsd.readthedocs.io/en/v4.0.0/) file format, which is the native file format for HOOMD-Blue.
-`msibi` includes a utility function that converts trajectory files gsd file format.
+`msibi` includes a utility function that converts trajectory files to the gsd file format.
[…]
