### 2025-10-01 702734f
Add JOSS paper status badge to README
Added a status badge for JOSS paper to README.
