### 2025-10-01 da840a9
Update install instructions required for running the tutorial examples
