+++ environment.yml
@@ -5,8 +5,8 @@ dependencies:
 - freud >=3.4
 - cmeutils >=1.4
 - gsd >=3.0
-- hoomd >=5.0
-- matplotlob
+- hoomd >=5.0,<5.4
+- matplotlib
 - python >=3.10,<=3.13
 - pandas
 - mdanalysis <3.0
