### 2025-10-23 3ef06fd
Add WSL test to CI workflow, add notes about using with Windows
