### 2025-11-09 #9244 @PabloIbannez
I appreciate the author’s clarification regarding the repository history. Given that sustained development has occurred in the repo the proposed authorship list looks reasonable to me.

**Minor comments** (paper text)

- On the nature of coarse-graining. I would modestly expand the framing: coarse-graining is not only a technical tool, but also a modeling philosophy about what to represent and what to omit. [Perspective: Coarse-grained models for biomolecularsystems, Noid, J. Chem. Phys. 139, 090901 (2013), https://doi.org/10.1063/1.4818908]

- MARTINI paragraph. The current sentence feels terse. If the intent is to contrast fixed bead types and transferability trade-offs, please elaborate with a balanced note and citations.

**Functionality and documentation**

- Tests: how to run and what they validate. The repository contains tests, but I could not find a clear “Testing” section describing:

1. The exact command to run (e.g., pytest -q from repo root after creating the conda env)
2. Brief mapping of what each test (or group of tests) is verifying.

- Examples and minimal data: The docs and README include code snippets and short examples, but they cannot be executed end-to-end without small, ready-to-use input trajectories (e.g., tiny GSD files). I recommend:

1. Shipping a minimal dataset (very small GSDs) or adding a script/notebook that generates them.
2. Providing a Google Colab (or similar) that creates the conda/mamba environment, downloads the minimal data, and runs one short MSIBI optimization to completion.
3. Linking that notebook from the README and the docs “Examples/Tutorials” page.

- Please add a short subsection in the README and/or docs (“Contributing”) with explicit steps.
