### 2025-11-10 34c38be
Wrap up validations into a single file to make it easier to call, update README to point to tutorial page in docs, run tests and validate
