### 2025-11-10 4aa67dc
Add contributing guidelines
