### 2025-11-10 45011ec
Paper updates and new citation
