### 2025-11-10 ctk3b/msibi#60 @chrisjonesBSU
Create repository for housing MSIBI example workflows
Following similar practice of other MoSDeF software, we should create a separate repository that contains Jupyter notebooks running examples from simple systems (re-creating Lennard Jones and harmonic potentials) to more complex cases (propane, polymers, etc..). We can render the Jupyter notebooks in the documentation for MSIBI as more are added.
