### 2025-11-10 #9244 @chrisjonesBSU
Thank you for the review and comments @PabloIbannez. Please see our responses and changes below.

> On the nature of coarse-graining. I would modestly expand the framing: coarse-graining is not only a technical tool, but also a modeling philosophy about what to represent and what to omit. [Perspective: Coarse-grained models for biomolecularsystems, Noid, J. Chem. Phys. 139, 090901 (2013), https://doi.org/10.1063/1.4818908]

This is a good point, especially for framing the statement of need for multiple kinds of software and methods for doing coarse-graining. The current word count is already on the high-end for a JOSS paper, so we tried to be as concise as possible in the coarse-graining section. We did take your suggestion and included the following (new text in bold):

_Coarse-graining (CG) is a commonly adopted solution to this challenge, as it reduces computational cost by grouping—or mapping—atoms into a single, larger bead [@ Joshi2021].
**Beyond being a technical tool to improve efficiency, coarse-graining also reflects modelling choices about which molecular features are essential to retain, and which can be simplified, which depends in large part on the materials being studied, the properties measured, and the researcher's objectives [@Noid2013].**
As a result, this approach introduces two challenges: first, the potential energy surface for a given chemistry and CG mapping is not known *a priori*, and second, as the mapping used is arbitrary, with multiple valid options, developing a single CG force field that is transferable across various mapping choices is not possible._

> MARTINI paragraph. The current sentence feels terse. If the intent is to contrast fixed bead types and transferability trade-offs, please elaborate with a balanced note and citations.

This is also a good point. Again, I think the terseness results from trying to limit word count for a JOSS paper. We reworded this to be more balanced. It now highlights the different approaches between MARTINI (top-down coarse-graining) and IBI/MSIBI (bottom-up coarse-graining). I included a citation to the Joshi 2021 reference used previously, which also contains a section on MARTINI that supports our statements here. The new text is:

_The MARTINI force field is a widely adopted and successful CG model that employs a top-down CG approach, where interactions are optimized to reproduce experimental properties rather than structural distributions from atomistic simulations, as done in IBI and MSIBI [@ Martini2007].This design provides excellent transferability and robustness across diverse systems, but its predefined bead types and fixed four-to-one mapping scheme limit adaptability to other resolutions and chemistries [@ Joshi2021]._

> Tests: how to run and what they validate. The repository contains tests, but I could not find a clear “Testing” section describing:
>     1. The exact command to run (e.g., pytest -q from repo root after creating the conda env)
>     2. Brief mapping of what each test (or group of tests) is verifying.

Thank you for pointing this out. I updated the README on the repository with instructions for running unit tests and accessing the simple validation tests included. In total, there are [68 separate unit tests](https://github.com/mosdef-hub/msibi/blob/main/msibi/tests/). They are separated by module (e.g., `test_state.py`, `test_corrections.py`, `test_forces.py` etc.), and each unit test is limited in scope with descriptive method names. We believe this should be sufficient for potential users or developers to understand what is being tested, and how, especially with so many tests.

Validation includes testing if `msibi` can recreate simple Lennard Jones (non-bonded) and harmonic (bonded) potentials. The README has been updated with both instructions on how to run this and what it validates. As of now, validating `msibi` on specific systems (e.g., small molecules, polymers, etc.) would be up to the researcher.

>     * Examples and minimal data: The docs and README include code snippets and short examples, but they cannot be executed end-to-end without small, ready-to-use input trajectories (e.g., tiny GSD files). I recommend:
>
>     1. Shipping a minimal dataset (very small GSDs) or adding a script/notebook that generates them.
>
>     2. Providing a Google Colab (or similar) that creates the conda/mamba environment, downloads the minimal data, and runs one short MSIBI optimization to completion.

We actually address this in the [Tutorials](https://msibi.readthedocs.io/en/latest/tutorials.html) section of the docs, although I see how that is not made clear in the README, or the Examples section of the documentation.

The repository contains a small set of .gsd files (the ones used in validation) that are also used in a tutorial example. We didn't want to package the files on anaconda, so the tutorials section has instructions on how to install `msibi` so that these files are accessible when running through the [tutorial](https://msibi.readthedocs.io/en/latest/tutorials/01-Introduction.html). We added information about this, with a link, in both the README and the Examples section of the documentation.

I agree, providing a Google Colab that sets up an environment and makes data sets accessible would be a great addition. What we plan on doing is creating a separate repository containing several Jupyter notebook examples that we, or anyone else, can add to with their own examples. A separate repository would help keep the source code clean while providing a library of examples we can link to. This follows what we've done with other MoSDeF packages (e.g., [mbuild](https://github.com/mosdef-hub/mbuild) and [mbuild-examples](https://github.com/mosdef-hub/mbuild-examples)). However, I am not positive I will be able to get to that in the time-frame of this review. I hope that the included examples, tutorial, and files, with updated instructions are sufficient for the time being. I opened an [issue](https://github.com/mosdef-hub/msibi/issues/60) on `msibi` to document this for the time being.

>     3. Linking that notebook from the README and the docs “Examples/Tutorials” page.

We added a link to the README that takes users to the [Tutorials](https://msibi.readthedocs.io/en/latest/tutorials.html) section of the docs with instructions on how to install `msibi` so that the simple data set is accessible.

>     * Please add a short subsection in the README and/or docs (“Contributing”) with explicit steps.

Thank you for the reminder! The latest PR in `msibi` added both a subsection to the README about contributing, and a CONTRIBUTING.MD file with specific guidelines and instructions for contributing.

Any changes to the paper, documentation or code mention above were made in https://github.com/mosdef-hub/msibi/pull/61
