### 2025-11-10 e9af5af
Fix validation test command in README.md
Update the path for validation tests in README.
