### 2025-11-10 34e97e5
Fix bibtex entry
