+++ paper/paper.bib
@@ -38,7 +38,7 @@
   volume={139},
   number={9},
   year={2013},
-  publisher={AIP Publishing}
+  publisher={AIP Publishing},
   doi={10.1063/1.4818908}
 }
 
