+++ paper/paper.md
@@ -54,7 +54,7 @@ IBI and MS-IBI are popular choices for deriving CG forces for polymers and biomo
 While these methods are frequently used, open-source software tools that provide an accessible and reproducible, end-to-end workflow for IBI and MS-IBI remain limited, especially for arbitrary mappings and multi-state systems.
 
 The MARTINI force field is a widely adopted and successful CG model that employs a top-down CG approach, where interactions are optimized to reproduce experimental properties rather than structural distributions from atomistic simulations, as done in IBI and MSIBI [@Martini2007].
-This design provides excellent transferability and robustness across diverse systems, but its predefined bead types and fixed four-to-one mapping scheme limit adaptability to other resolutions and chemistries [@Joshi2021].
+This design provides excellent transferability and robustness across diverse systems, but its predefined bead types and fixed mapping rules limit flexibility for arbitrary resolutions or novel chemistries [@Joshi2021].
 The Versatile Object-oriented Toolkit for Coarse-graining Applications (VOTCA) offers a robust implementation of IBI—among several other features—and is also widely used in the community [@Baumeier2024].
 However, its workflow relies on manual management of multiple input files and bash operations, which can introduce operational complexity that reduces reproducibility and usability [@Cummings_2020; @Jankowski2019].
 Additionally, VOTCA's implementation of IBI does not natively support inclusion and weighting of multiple state points, as implemented in the MS-IBI method.
