### 2025-11-26 2202fd8
Update doc strings, change default smoothing param values
