### 2025-12-09 5fc5415
bump ruff version and precommit fixes
