+++ .pre-commit-config.yaml
@@ -10,7 +10,7 @@ ci:
 repos:
 - repo: https://github.com/astral-sh/ruff-pre-commit
   # Ruff version.
-  rev: v0.12.3
+  rev: v0.14.8
   hooks:
     # Run the linter.
     - id: ruff
