### 2026-01-21 #9244 @chrisjonesBSU
Hello all @jedbrown @valsson @danielskatz. I'm just checking in again. I realize that the beginning of a new semester may be a very busy time for @jedbrown and @valsson, also, just from looking around other JOSS submissions, I see that the demand on reviewers is quite high as the moment. We aren't in a rush for any particular reason, but I'll keep pinging to hopefully get this closed up soon! Hopefully with one thorough review already addressed, the second one won't take too much time.
