### 2026-01-28 #9244 @ajayi-joseph
Hi I'm a registered JOSS reviewer and would be happy to help with this review if an additional reviewer would be useful. I have 7+ years of software engineering experience and maintain open source packages.
