### 2026-03-02 #9244 @valsson
@chrisjonesBSU I sincerely apologize again for the delay in performing the review! Here is the first part of my review.

### Paper

The current version of the paper looks great! @PabloIbannez comments addressed most of the issues. I have no issues with the paper. It is clear and describes the high-level functionality and purpose of the software clearly.

### Tutorials

The current tutorial, "MSIBI: Simple Tutorial", is sufficient for this JOSS submission, as it covers the main aspects of the code and how it works. However, the current tutorial lacks in showing the users how the code is used in more practical/real cases and workflows. Thus, I agree with @PabloIbannez that it would be good to add/create a dedicated examples/notebooks repository showing practical/real workflows. The authors are already planning this (see https://github.com/mosdef-hub/msibi/issues/60). Here, it would be helpful to incorporate the IBI theory into the tutorial. In these tutorials, it would be good to show how to define the CG mappings and how to use the resulting CG potentials in supported codes. The ACS Journal of Physical Chemistry A/B/C has a Tutorial article type (see e.g., [DOI:10.1021/acs.jpcb.5c04333](https://pubs.acs.org/doi/10.1021/acs.jpcb.5c04333) and [DOI:10.1021/acs.jpcb.4c04901](https://pubs.acs.org/doi/10.1021/acs.jpcb.4c04901)), so the authors might consider writing their tutorials into an article so it counts as a publication.

### Code, Repository, and Documentations

I just have a few comments regarding the code

I have managed to install the code via the Github repo, but I observed some issues. As I observed some issues related to conda, I should note that I am using miniforge and conda 26.1.1.

#### 1
The path to the validation script in the `Readme.md` file is wrong; it is given as `python msibi/tests/validation/validation.py` but it is `python  msibi/tests/validation/validate.py`

#### 2
The `Readme.md` file uses `mamba` while the Documentation/Tutorial uses `conda`; it would be good to make that consistent.

#### 3
When I tried to create a conda environment using `conda env create -f environment.yml python=3.13 jupyter` as described in the tutorial page (https://msibi.readthedocs.io/en/latest/tutorials.htm), I got an error message:

```
> conda env create -f environment.yml python=3.13 jupyter
usage: conda [-h] [-v] [--no-plugins] [-V] COMMAND ...
conda: error: unrecognized arguments: python=3.13 jupyter
> conda env create -f environment.yml python=3.13 jupyter
```

The `conda env create -f environment.yml ` command is also different from the ones listed in the installation page of the documentation (https://msibi.readthedocs.io/en/latest/installation.html).

Using `conda env create -f environment.yml` worked fine.

#### 4

When I use the `conda env create -f environment.yml` command, I get the following warning:

```
Warning: you have pip-installed dependencies in your environment file, but you do not list pip itself as one of your conda dependencies.  Conda may not use the correct pip to install your packages, and they may end up in the wrong place.  Please add an explicit pip dependency.  I'm adding one for you, but still nagging you.
```

#### 5
I can run `pytest` as described in `Readme.md`, and I get the following output:
`============================================ 68 passed, 22 warnings in 5.88s ========`
But it is not clear to me whether I am expected to get these 22 warnings. Is that something that the user can ignore? Or does this reflect some issues with my installation?

#### 6
Typo in tutorial, `damping potnetial corrections`

#### 7

In the tutorial, I understand that this is the code to specify where the files needed for the tutorial are located:
```
with importlib.resources.path(
    "msibi.tests.validation", "target_data", "chains"
) as target_dir:
    target_data_dir = target_dir
```
However, this part of the code might feel a bit cryptic, so it would be good to explain what is being done here, and where the files are located.

### Running the code

At the moment, I have some issues running the code and reproducing the tutorial, but I will list them in a separate comment.

Once these issues have been resolved and I have successfully reproduced the tutorial results, I will finalize my review.
