### 2026-03-02 #9244 @valsson
I have managed to install the code via the GitHub repo as instructed in the tutorial part, but I have some issues that I haven't been able to resolve quickly, so I haven't been able to run through the tutorial.

#### Unit tests

Running the unit tests with `pytest` gives the following input:
```
================================================== test session starts ===================================================
rootdir: /home/ov0060/programs/msibi
configfile: pyproject.toml
plugins: anyio-4.12.1
collected 68 items

msibi/tests/test_conversion.py ......                                                                              [  8%]
msibi/tests/test_corrections.py ......                                                                             [ 17%]
msibi/tests/test_forces.py ..................................                                                      [ 67%]
msibi/tests/test_optimize.py .........                                                                             [ 80%]
msibi/tests/test_potentials.py .                                                                                   [ 82%]
msibi/tests/test_state.py ..........                                                                               [ 97%]
msibi/tests/test_utils.py ..                                                                                       [100%]

==================================================== warnings summary ====================================================
msibi/tests/test_conversion.py::TestConversion::test_no_file
  <frozen importlib._bootstrap>:488: DeprecationWarning: builtin type SwigPyPacked has no __module__ attribute

msibi/tests/test_conversion.py::TestConversion::test_no_file
  <frozen importlib._bootstrap>:488: DeprecationWarning: builtin type SwigPyObject has no __module__ attribute

msibi/tests/test_conversion.py::TestConversion::test_no_file
  <frozen importlib._bootstrap>:488: DeprecationWarning: builtin type swigvarlink has no __module__ attribute

msibi/tests/test_conversion.py::TestConversion::test_lammps_dump
  /home/ov0060/.conda/envs/msibi-repo/lib/python3.13/site-packages/MDAnalysis/coordinates/LAMMPS.py:749: UserWarning: Reader has no dt information, set to 1.0 ps
    ts.data["time"] = step_num * ts.dt

msibi/tests/test_conversion.py::TestConversion::test_lammps_dcd
msibi/tests/test_conversion.py::TestConversion::test_lammps_dcd
  /home/ov0060/.conda/envs/msibi-repo/lib/python3.13/site-packages/MDAnalysis/coordinates/DCD.py:171: DeprecationWarning: DCDReader currently makes independent timesteps by copying self.ts while other readers update self.ts inplace. This behavior will be changed in 3.0 to be the same as other readers. Read more at https://github.com/MDAnalysis/mdanalysis/issues/3889 to learn if this change in behavior might affect you.
    warnings.warn("DCDReader currently makes independent timesteps"

msibi/tests/test_forces.py::TestForce::test_fit_scores
msibi/tests/test_optimize.py::TestMSIBI::test_run
msibi/tests/test_optimize.py::TestMSIBI::test_run_with_static_force
  /home/ov0060/.conda/envs/msibi-repo/lib/python3.13/site-packages/cmeutils/structure.py:264: UserWarning: There are bond lengths that fall outside of your set l_min and l_max range. You may want to adjust this range to include all bond lengths.
    warnings.warn(

msibi/tests/test_forces.py::TestForce::test_fit_scores
msibi/tests/test_optimize.py::TestMSIBI::test_run
msibi/tests/test_optimize.py::TestMSIBI::test_run_with_static_force
msibi/tests/test_optimize.py::TestMSIBI::test_optimize_pairs_linear_alpha
msibi/tests/test_optimize.py::TestMSIBI::test_run_with_all_forces
  /home/ov0060/programs/msibi/msibi/forces.py:700: RuntimeWarning: divide by zero encountered in divide
    state.kT * np.log(current_dist[:, 1] / target_dist[:, 1]) / N

msibi/tests/test_forces.py::TestForce::test_fit_scores
msibi/tests/test_optimize.py::TestMSIBI::test_run
msibi/tests/test_optimize.py::TestMSIBI::test_run_with_static_force
msibi/tests/test_optimize.py::TestMSIBI::test_optimize_pairs_linear_alpha
msibi/tests/test_optimize.py::TestMSIBI::test_run_with_all_forces
  /home/ov0060/programs/msibi/msibi/forces.py:700: RuntimeWarning: invalid value encountered in divide
    state.kT * np.log(current_dist[:, 1] / target_dist[:, 1]) / N

msibi/tests/test_forces.py::TestForce::test_fit_scores
msibi/tests/test_optimize.py::TestMSIBI::test_run
msibi/tests/test_optimize.py::TestMSIBI::test_run_with_static_force
  /home/ov0060/programs/msibi/msibi/forces.py:700: RuntimeWarning: divide by zero encountered in log
    state.kT * np.log(current_dist[:, 1] / target_dist[:, 1]) / N

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
============================================ 68 passed, 22 warnings in 5.88s =============================================
```

Is this something that I need to worry about?

#### Validation tests

Also, running the validation tests gives me the following error:
```
(msibi-repo) ~/programs/msibi$ python msibi/tests/validation/validate.py
---Optimization: 1 of 12---
Starting simulation 0 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf1abf230>
Finished simulation 0 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 0, Fit score:0.286560

---Optimization: 2 of 12---
Starting simulation 1 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf1a5fed0>
Finished simulation 1 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 1, Fit score:0.673942

---Optimization: 3 of 12---
Starting simulation 2 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbef604910>
Finished simulation 2 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 2, Fit score:0.855236

---Optimization: 4 of 12---
Starting simulation 3 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf1af7e10>
Finished simulation 3 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 3, Fit score:0.903050

---Optimization: 5 of 12---
Starting simulation 4 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf1af7490>
Finished simulation 4 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 4, Fit score:0.962131

---Optimization: 6 of 12---
Starting simulation 5 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf10c32f0>
Finished simulation 5 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 5, Fit score:0.930291

---Optimization: 7 of 12---
Starting simulation 6 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbef65caf0>
Finished simulation 6 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 6, Fit score:0.959459

---Optimization: 8 of 12---
Starting simulation 7 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbef65c270>
Finished simulation 7 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 7, Fit score:0.942919

---Optimization: 9 of 12---
Starting simulation 8 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf1b02750>
Finished simulation 8 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 8, Fit score:0.975978

---Optimization: 10 of 12---
Starting simulation 9 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf1b01d50>
Finished simulation 9 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 9, Fit score:0.955860

---Optimization: 11 of 12---
Starting simulation 10 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf10df890>
Finished simulation 10 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 10, Fit score:0.928085

---Optimization: 12 of 12---
Starting simulation 11 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xedbf10dee40>
Finished simulation 11 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 11, Fit score:0.974297

Traceback (most recent call last):
  File "/home/ov0060/programs/msibi/msibi/tests/validation/validate.py", line 197, in <module>
    validate_bonds()
    ~~~~~~~~~~~~~~^^
  File "/home/ov0060/programs/msibi/msibi/tests/validation/validate.py", line 57, in validate_bonds
    optimizer.run_optimization(n_iterations=12, n_steps=5e5, backup_trajectories=False)
    ~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/ov0060/programs/msibi/msibi/optimize.py", line 201, in run_optimization
    self._update_potentials()
    ~~~~~~~~~~~~~~~~~~~~~~~^^
  File "/home/ov0060/programs/msibi/msibi/optimize.py", line 287, in _update_potentials
    force._update_potential()
    ~~~~~~~~~~~~~~~~~~~~~~~^^
  File "/home/ov0060/programs/msibi/msibi/forces.py", line 715, in _update_potential
    self._potential, head_cut, tail_cut, real_indices = bonded_corrections(
                                                        ~~~~~~~~~~~~~~~~~~^
        x=self.x_range,
        ^^^^^^^^^^^^^^^
    ...<6 lines>...
        tail_correction_func=self.correction_form,
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    )
    ^
  File "/home/ov0060/programs/msibi/msibi/utils/corrections.py", line 105, in bonded_corrections
    popt_head, pcov_head = curve_fit(
                           ~~~~~~~~~^
        f=head_correction_func,
        ^^^^^^^^^^^^^^^^^^^^^^^
    ...<2 lines>...
        maxfev=maxfev,
        ^^^^^^^^^^^^^^
    )
    ^
  File "/home/ov0060/.conda/envs/msibi-repo/lib/python3.13/site-packages/scipy/optimize/_minpack_py.py", line 1026, in curve_fit
    raise RuntimeError("Optimal parameters not found: " + errmsg)
RuntimeError: Optimal parameters not found: Number of calls to function has reached maxfev = 1000.
```

#### Tutorial

I hit a similar issue in the tutorial at https://msibi.readthedocs.io/en/latest/tutorials/01-Introduction.html for this part (cell [8]):
```
optimizer.run_optimization(n_iterations=7, n_steps=1e5, backup_trajectories=False)
bond.smooth_potential()
```
where I get the following error:
```
---Optimization: 1 of 7---
Starting simulation 0 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5
Running on device <hoomd.device.CPU object at 0xdbb43bb8d70>
Finished simulation 0 for state <class 'msibi.state.State'>; Name: single-chain; kT: 4.0; Alpha0: 0.5

Force: A-A, State: single-chain, Iteration: 0, Fit score:0.325526

RuntimeError                              Traceback (most recent call last)
Cell In[5], line 1
----> 1 optimizer.run_optimization(n_iterations=7, n_steps=1e5, backup_trajectories=False)
      2 bond.smooth_potential()

File ~/programs/msibi/msibi/optimize.py:201, in MSIBI.run_optimization(self, n_steps, n_iterations, backup_trajectories, _dir)
    187 for state in self.states:
    188     state._run_simulation(
    189         n_steps=n_steps,
    190         forces=forces,
   (...)    199         backup_trajectories=backup_trajectories,
    200     )
--> 201 self._update_potentials()
    202 self.n_iterations += 1

File ~/programs/msibi/msibi/optimize.py:287, in MSIBI._update_potentials(self)
    285 for force in self._optimize_forces:
    286     self._recompute_distribution(force)
--> 287     force._update_potential()

File ~/programs/msibi/msibi/forces.py:715, in Force._update_potential(self)
    704     self._potential, head_cut, tail_cut, real_indices = pair_corrections(
    705         x=self.x_range,
    706         V=self.potential,
   (...)    712         head_correction_func=self.correction_form,
    713     )
    714 else:  # Bonded force, use correction form on both head and tail of potential
--> 715     self._potential, head_cut, tail_cut, real_indices = bonded_corrections(
    716         x=self.x_range,
    717         V=self.potential,
    718         smoothing_window=self.smoothing_window,
    719         smoothing_order=self.smoothing_order,
    720         fit_window_size=self.correction_fit_window,
    721         maxfev=self.maxfev,
    722         head_correction_func=self.correction_form,
    723         tail_correction_func=self.correction_form,
    724     )
    725 # Store all versions of final potential (actually learned, and extrapolated)
    726 self.potential_history.append(np.copy(self.potential))

File ~/programs/msibi/msibi/utils/corrections.py:105, in bonded_corrections(x, V, smoothing_window, smoothing_order, fit_window_size, maxfev, head_correction_func, tail_correction_func)
    103 x_head_pivot = x_real[fit_window_size - 1]
    104 x_head_fit = _shift_x(x_real[:fit_window_size], origin=x_head_pivot)
--> 105 popt_head, pcov_head = curve_fit(
    106     f=head_correction_func,
    107     xdata=x_head_fit,
    108     ydata=v_real[:fit_window_size],
    109     maxfev=maxfev,
    110 )
    111 x_head_missing = _shift_x(x[:head_start], origin=x_head_pivot)
    112 # Apply these parameters to the x-range where we are missing data

File ~/.conda/envs/msibi-repo/lib/python3.13/site-packages/scipy/optimize/_minpack_py.py:1026, in curve_fit(f, xdata, ydata, p0, sigma, absolute_sigma, check_finite, bounds, method, jac, full_output, nan_policy, **kwargs)
   1024     cost = np.sum(infodict['fvec'] ** 2)
   1025     if ier not in [1, 2, 3, 4]:
-> 1026         raise RuntimeError("Optimal parameters not found: " + errmsg)
   1027 else:
   1028     # Rename maxfev (leastsq) to max_nfev (least_squares), if specified.
   1029     if 'max_nfev' not in kwargs:

RuntimeError: Optimal parameters not found: Number of calls to function has reached maxfev = 1000.
```

Do you have an idea what the issue could be?

#### Python packages

I list the installed Python packages in case that helps.

```
(msibi-repo) ov0060@hekla:~/programs/msibi/tutorial-test$ conda list
# packages in environment at /home/ov0060/.conda/envs/msibi-repo:
#
# Name                           Version          Build                      Channel
_openmp_mutex                    4.5              20_gnu                     conda-forge
_python_abi3_support             1.0              hd8ed1ab_2                 conda-forge
alsa-lib                         1.2.15.3         hb03c661_0                 conda-forge
annotated-types                  0.7.0            pyhd8ed1ab_1               conda-forge
anyio                            4.12.1           pyhcf101f3_0               conda-forge
argon2-cffi                      25.1.0           pyhd8ed1ab_0               conda-forge
argon2-cffi-bindings             25.1.0           py313h07c4f96_2            conda-forge
arrow                            1.4.0            pyhcf101f3_0               conda-forge
asttokens                        3.0.1            pyhd8ed1ab_0               conda-forge
async-lru                        2.2.0            pyhcf101f3_0               conda-forge
attr                             2.5.2            h39aace5_0                 conda-forge
attrs                            25.4.0           pyhcf101f3_1               conda-forge
babel                            2.18.0           pyhcf101f3_0               conda-forge
backports.zstd                   1.3.0            py313h18e8e13_0            conda-forge
beautifulsoup4                   4.14.3           pyha770c72_0               conda-forge
bleach                           6.3.0            pyhcf101f3_1               conda-forge
bleach-with-css                  6.3.0            hbca2aae_1                 conda-forge
blosc                            1.21.6           he440d0b_1                 conda-forge
boltons                          25.0.0           pyhd8ed1ab_0               conda-forge
brotli                           1.2.0            hed03a55_1                 conda-forge
brotli-bin                       1.2.0            hb03c661_1                 conda-forge
brotli-python                    1.2.0            py313hf159716_1            conda-forge
build                            1.4.0            pypi_0                     pypi
bzip2                            1.0.8            hda65f42_9                 conda-forge
c-ares                           1.34.6           hb03c661_0                 conda-forge
c-blosc2                         2.23.0           hc31b594_0                 conda-forge
ca-certificates                  2026.2.25        hbd8a1cb_0                 conda-forge
cached-property                  1.5.2            hd8ed1ab_1                 conda-forge
cached_property                  1.5.2            pyha770c72_1               conda-forge
cairo                            1.18.4           he90730b_1                 conda-forge
certifi                          2026.2.25        pyhd8ed1ab_0               conda-forge
cffi                             2.0.0            py313hf46b229_1            conda-forge
cftime                           1.6.5            py313h29aa505_1            conda-forge
charset-normalizer               3.4.4            pyhd8ed1ab_0               conda-forge
cmeutils                         1.4.0            pyhd8ed1ab_0               conda-forge
colorama                         0.4.6            pyhd8ed1ab_1               conda-forge
comm                             0.2.3            pyhe01879c_0               conda-forge
contourpy                        1.3.3            py313hc8edb43_4            conda-forge
cpython                          3.13.12          py313hd8ed1ab_100          conda-forge
cuda-cccl_linux-64               12.9.27          ha770c72_0                 conda-forge
cuda-crt-dev_linux-64            12.9.86          ha770c72_2                 conda-forge
cuda-crt-tools                   12.9.86          ha770c72_2                 conda-forge
cuda-cudart                      12.9.79          h5888daf_0                 conda-forge
cuda-cudart-dev_linux-64         12.9.79          h3f2d84a_0                 conda-forge
cuda-cudart-static_linux-64      12.9.79          h3f2d84a_0                 conda-forge
cuda-cudart_linux-64             12.9.79          h3f2d84a_0                 conda-forge
cuda-cupti                       12.9.79          h676940d_1                 conda-forge
cuda-cupti-dev                   12.9.79          h676940d_1                 conda-forge
cuda-nvcc-tools                  12.9.86          he02047a_2                 conda-forge
cuda-nvrtc                       12.9.86          hecca717_1                 conda-forge
cuda-nvtx                        12.9.79          hecca717_1                 conda-forge
cuda-nvvm-tools                  12.9.86          h4bc722e_2                 conda-forge
[…]
