### 2026-03-03 eb177ff
Fix path to validate.py in README
