### 2026-03-03 fc7d496
Replace mamba with conda in README
