### 2026-03-03 4165514
Add note about states dir and validate.py
