### 2026-03-03 3238dd7
Add pip to both .yml files
