+++ environment-dev.yml
@@ -15,5 +15,6 @@ dependencies:
 - pytest
 - pytest-cov
 - pre-commit
+- pip
 - pip:
     - build
