### 2026-03-03 54b9325
Use better sampling parameters in validate.py
