### 2026-03-03 be8aa8d
Improve error message, ensure consistency between tutorial and validate.py
