+++ msibi/tests/validation/validate.py
@@ -29,7 +29,8 @@ def validate_bonds():
         thermostat_kwargs={"tau": 0.1},
         method_kwargs={},
         dt=0.0001,
-        gsd_period=5000,
+        gsd_period=500,
+        seed=50,
     )
 
     state = State(
@@ -48,13 +49,14 @@ def validate_bonds():
         nbins=60,
         smoothing_window=5,
         correction_fit_window=7,
+        smoothing_order=1,
     )
     bond.set_polynomial(x_min=0.0, x_max=2.5, x0=1.3, k2=200, k3=0, k4=0)
 
     optimizer.add_state(state)
     optimizer.add_force(bond)
 
-    optimizer.run_optimization(n_iterations=7, n_steps=1e6, backup_trajectories=False)
+    optimizer.run_optimization(n_iterations=10, n_steps=1e6, backup_trajectories=False)
     bond.smooth_potential()
 
     scores = bond._states[state]["f_fit"]
@@ -63,11 +65,11 @@ def validate_bonds():
     # Target data simulations used k = 500 and x0 = 1.1
     params, params_covariance = curve_fit(harmonic, bond.x_range, bond.potential)
     k_fit, x0_fit = params
-    assert np.allclose(k_fit, 500, atol=20)
-    assert np.allclose(x0_fit, 1.1, atol=0.05)
 
     print("Finished validating bonds.")
     print(f"Fit score = {scores[-1]}, k fit = {k_fit}, x0 fit = {x0_fit}")
+    assert np.allclose(k_fit, 500, atol=20)
+    assert np.allclose(x0_fit, 1.1, atol=0.05)
 
     # Clean-up
     shutil.rmtree(os.path.join(_dir, "states"))
@@ -81,7 +83,8 @@ def validate_angles():
         thermostat_kwargs={"tau": 0.1},
         method_kwargs={},
         dt=0.0001,
-        gsd_period=5000,
+        gsd_period=500,
+        seed=50,
     )
 
     state = State(
@@ -104,6 +107,7 @@ def validate_angles():
         nbins=60,
         smoothing_window=5,
         correction_fit_window=7,
+        smoothing_order=1
     )
     angle.set_polynomial(x_min=0.0, x_max=np.pi, x0=2.3, k2=80, k3=0, k4=0)
 
@@ -128,7 +132,7 @@ def validate_angles():
 
     print("Finished")
     print(f"Fit score = {scores[-1]}, k fit = {k_fit}, x0 fit = {x0_fit}")
-    assert np.allclose(k_fit, 250, atol=26)
+    assert np.allclose(k_fit, 250, atol=30)
     assert np.allclose(x0_fit, 2.0, atol=0.05)
 
     # Clean-up
@@ -143,7 +147,8 @@ def validate_pairs():
         thermostat_kwargs={"tau": 0.1},
         method_kwargs={},
         dt=0.0001,
-        gsd_period=5000,
+        gsd_period=1000,
+        seed=50,
     )
 
     kT = 3.0
@@ -168,7 +173,7 @@ def validate_pairs():
         optimize=True,
         r_cut=2.5,
         nbins=80,
-        smoothing_window=15,
+        smoothing_window=11,
         r_switch=1.5,
         correction_fit_window=7,
     )
