### 2026-03-04 #9244 @valsson
I have managed to successfully run the tutorial on my Apple laptop. The previous case where I had issues was on my Ubuntu 24.04LTS workstation.

I get slightly different results from the results shown on the tutorial page, for example, see here

Are these differences something that can be expected? Perhaps due to stochasticness in the fitting or simulation process?
