### 2026-03-04 #9244 @valsson
I noticed some issues with the validation script when running it on my Apple laptop.

First, running it from the root folder,
```
python msibi/tests/validation/validate.py
```
I get the following error message
```
  File "/Users/omar.valsson/programs/msibi/msibi/tests/validation/validate.py", line 197, in <module>
    validate_bonds()
    ~~~~~~~~~~~~~~^^
  File "/Users/omar.valsson/programs/msibi/msibi/tests/validation/validate.py", line 73, in validate_bonds
    shutil.rmtree(os.path.join(_dir, "states"))
    ~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/opt/homebrew/Caskroom/miniforge/base/envs/msibi/lib/python3.13/shutil.py", line 763, in rmtree
    _rmtree_safe_fd(stack, onexc)
    ~~~~~~~~~~~~~~~^^^^^^^^^^^^^^
  File "/opt/homebrew/Caskroom/miniforge/base/envs/msibi/lib/python3.13/shutil.py", line 704, in _rmtree_safe_fd
    onexc(func, path, err)
    ~~~~~^^^^^^^^^^^^^^^^^
  File "/opt/homebrew/Caskroom/miniforge/base/envs/msibi/lib/python3.13/shutil.py", line 665, in _rmtree_safe_fd
    orig_st = os.lstat(name, dir_fd=dirfd)
FileNotFoundError: [Errno 2] No such file or directory: '/Users/omar.valsson/programs/msibi/msibi/tests/validation/states'
```
So it doesn't work to run it from the root folder.

Going to the validation folder and running the `validate.py` script there works
```
> % cd msibi/tests/validation
>% python validate.py
```
However, in that case, I get the following error:
```
Force: A-A-A, State: single-chain, Iteration: 11, Fit score:0.967205

Finished
Fit score = 0.9672045313913085, k fit = 278.40966416811347, x0 fit = 1.9803301162050682
Traceback (most recent call last):
  File "/Users/omar.valsson/programs/msibi/msibi/tests/validation/validate.py", line 198, in <module>
    validate_angles()
    ~~~~~~~~~~~~~~~^^
  File "/Users/omar.valsson/programs/msibi/msibi/tests/validation/validate.py", line 131, in validate_angles
    assert np.allclose(k_fit, 250, atol=26)
           ~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^
AssertionError
```
so the `np.allclose(k_fit, 250, atol=26)` assert check fails. I am not sure what the issue is here. It might be better to give the user some more verbose and clearer error message if the assert check fails.
