### 2026-03-04 6ba1ddd
Update hoomd pin in yml files
