+++ environment-dev.yml
@@ -5,7 +5,7 @@ dependencies:
 - cmeutils >=1.4
 - freud >=3.4
 - gsd >=3.0
-- hoomd >=5.0,<5.4
+- hoomd >=5..4
 - python >=3.10,<=3.13
 - pandas
 - numpy >=2.0,<2.3
