### 2026-03-04 a876aa3
Use same params between tutorial and validation
