### 2026-03-04 8c0054f
Fix pair head correction when r_switch is None
