### 2026-03-04 a06ac78
update instructions to run validate.py
