### 2026-03-04 #9244 @chrisjonesBSU
Thanks @valsson. I'll respond to your comments below.

> ### Code, Repository, and Documentations
>
> I just have a few comments regarding the code
>
> I have managed to install the code via the Github repo, but I observed some issues. As I observed some issues related to conda, I should note that I am using miniforge and conda 26.1.1.
> #### 1
>
> The path to the validation script in the `Readme.md` file is wrong; it is given as `python msibi/tests/validation/validation.py` but it is `python msibi/tests/validation/validate.py`

Thanks for catching this. I corrected the path.

> #### 2
> The `Readme.md` file uses `mamba` while the Documentation/Tutorial uses `conda`; it would be good to make that consistent.

I updated the README file to use `conda`, so `conda` should appear throughout the README, documentation and tutorial now.

> #### 3
> When I tried to create a conda environment using `conda env create -f environment.yml python=3.13 jupyter` as described in the tutorial page ([msibi.readthedocs.io/en/latest/tutorials.htm](https://msibi.readthedocs.io/en/latest/tutorials.htm)), I got an error message:

Good catch. I think this arises from differences between `mamba` and `conda`.
I updated this by removing the python version specification (it's not really needed) and moving `conda install jupyter` to its own command after running `conda activate msibi`

> The `conda env create -f environment.yml ` command is also different from the ones listed in the installation page of the documentation ([msibi.readthedocs.io/en/latest/installation.html](https://msibi.readthedocs.io/en/latest/installation.html)).
>
> Using `conda env create -f environment.yml` worked fine.

It is purposely different in the tutorials section to make sure jupyter is installed, as It's not really a run-time dependency for msibi.

> #### 4
> When I use the `conda env create -f environment.yml` command, I get the following warning:
>
> ```
> Warning: you have pip-installed dependencies in your environment file, but you do not list pip itself as one of your conda dependencies.  Conda may not use the correct pip to install your packages, and they may end up in the wrong place.  Please add an explicit pip dependency.  I'm adding one for you, but still nagging you.
> ```

I added `pip` to both `.yml` files to avoid this warning.

> #### 5
> I have managed to install the code via the GitHub repo as instructed in the tutorial part, but I have some issues that I haven't been able to resolve quickly, so I haven't been able to run through the tutorial.
> #### Unit tests
>
> Running the unit tests with `pytest` gives the following input:
>
> ```
> -- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
> ============================================ 68 passed, 22 warnings in 5.88s =============================================
> ```
>
> Is this something that I need to worry about?

No, these shouldn't be worried about, and I'd say they are unavoidable. I checked them and I think the warnings fall under 3 categories.

1) Deprecation warnings for python built-in packages. These will be addressed as needed when I bump the python versions used in testing and continuous-integration.

2) MDAnalysis (which we use for converting non .gsd trajectories to .gsd files) warning about a missing time step in a trajectory file, which is not an issue. That information isn't carried forward into the MSIBI run, as it is set manually by the user.

3) Divide by zero warnings. This is expected for the MSIBI method where the distribution contains values of 0.

> #### 6
> Typo in tutorial, `damping potnetial corrections`

This typo has been fixed in the tutorial.

> #### 7
> In the tutorial, I understand that this is the code to specify where the files needed for the tutorial are located:
>
> ```
> with importlib.resources.path(
>     "msibi.tests.validation", "target_data", "chains"
> ) as target_dir:
>     target_data_dir = target_dir
> ```
>
> However, this part of the code might feel a bit cryptic, so it would be good to explain what is being done here, and where the files are located.

Good point. I added some in-line comments to explain that this bit of code is needed only for running the tutorial as it defines the path to the files used for the tutorial only.

> #### Errors in Validation tests and Tutorial

Ok, this seems like a similar issue with the first reviewer, and is a pain-point in general. It's surprising that the tutorial isn't successfully running again, but I can reproduce your issue. What's happening is that in order to apply corrections (i.e., extrapolations to the potential) in the poorly sampled (or not sampled at all) regions, we try to fit the correction function (harmonic, linear, exponential) to a region of the learned potential. scipy's optimize is failing and throwing this long, unhelpful error message.

I moved this bit of code under a `try/except` statement to capture it and throw an error message that is more helpful. Also, it prints the following to help guide the user.

```
There was an error fitting the existing potential against the head/tail correction function.
This could indicate poor sampling, or a poor choice of fitting parameters.

Try adjusting parameters that impact sampling such as gsd_period in msibi.MSIBI,
sampling_stride and n_frames in msibi.State, and n_steps in MSIBI.run_optimization.
You might adjust the fitting and smoothing related parameters.
See smoothing_window, correction_fit_window, smoothing_order, maxfev in msibi.Force.
```

I also added another note in the tutorials related to possible sensitivity to fitting during these corrections.

Also, while investigating this, I realized that the `seed` parameter defined in the management class (`optimize.MSIBI`) wasn't being passed to HOOMD for the iterative simulations, which is now fixed.

Ultimately, I'd like to re-work the fitting methods to a more robust approach, but at the moment I won't be able to get to that. From my experience with system sizes beyond these small toy-systems (that need to run relatively quickly in testing and tutorials), these kinds of issues aren't as sensitive as they appear to be here. Also, there are several parameters that ultimately impact the head/tail correction and fitting. Ranging from the target trajectory itself, to the number of bins used in distributions and potentials, length and frequency of sampling in MSIBI trajectories, etc. So, that is to say, these are all essentially hyperparameters that the user will have to spend time tuning and understanding themselves for their own systems. I added a note to the tutorial about this after the previous review, but I'll add a section to the documentation as well. This would be a great topic to cover in a tutorial once I'm able to dedicate time to creating a more comprehensive set of tutorials.

As far as specifically addressing the errors.  I changed the `smoothing_order` to 1 for the bond and angle optimization runs, lowered the `alpha0` parameter, and ran more iterations at decreased steps per iteration. Both `validate.py` and the notebook run for me after testing it on an M4 MacBook, and a x84 Linux laptop (testing both CPU and GPU HOOMD).

Also, from my last time adjusting the tutorial notebook, there were some differences between the parameters used in `validate.py` and the notebook. They should essentially be running the same procedure now. `validate.py` just includes assertion checks against distribution matching scores whereas the tutorial plots the comparison to show the fit.

To re-test with the included changes, please first run a `git pull` and rebuild the msibi environment

```
conda env create -f environment.yml
conda activate msibi
conda install jupyter
pip install .
```

These changes were added in https://github.com/mosdef-hub/msibi/pull/67
