### 2026-03-04 #9244 @chrisjonesBSU
Oops, it looks like I started drafting my response before your last replies, which I didn't see before submitting it. Please see my comment above, it addresses issues with the tutorial and validation testing in general.
