### 2026-03-04 #9244 @valsson
Hello @chrisjonesBSU, thank you for the response and changes!

I have tested the updated code and can run the validation script on both my Apple laptop and Linux workstation without issues.

I have also run through the tutorial on both, and it now runs fine and without issues. I am getting slightly different results for some of the figures, but I guess that is just due to differences in the random numbers, which can be expected for MD simulations and fitting, so that is fine.

I noticed one typo in the tutorial: "...automatically calcualted with the given..".

Now, I think this paper is good to go. I have no further comments or remarks. I have checked off all the items in my review checklist above.

I apologize again for the delay in doing the review!
