### 2026-03-09 fe81a0d
Add ability to ignore states for a Foce object
