+++ msibi/forces.py
@@ -1,6 +1,6 @@
 import os
 import warnings
-from typing import Callable, Optional, Union
+from typing import Callable, Optional, Union, List
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -101,6 +101,7 @@ class Force:
         correction_fit_window: Optional[int] = None,
         maxfev: Optional[int] = 3000,
         correction_form: Optional[Callable] = None,
+        ignore_states: List = None 
     ):
         if optimize and not nbins or optimize and nbins <= 0:
             raise ValueError(
@@ -113,6 +114,10 @@ class Force:
         self.correction_fit_window = correction_fit_window if optimize else None
         self.maxfev = int(maxfev)
         self.correction_form = correction_form
+        if ignore_states is None:
+            self.ignore_states = []
+        else:
+            self.ignore_states = ignore_states
         self._smoothing_window = smoothing_window if optimize else None
         self._smoothing_order = smoothing_order if optimize else None
         self.format = None
@@ -690,15 +695,19 @@ class Force:
 
     def _update_potential(self) -> None:
         """Compare distributions and update potential via Boltzmann Inversion."""
+        state_count = 0 
         for state in self._states:
+            if state in self.ignore_states:
+                continue
             current_dist = self._states[state]["current_distribution"]
             target_dist = self._states[state]["target_distribution"]
             self._states[state]["distribution_history"].append(current_dist)
-            N = len(self._states)
             alpha_array = state.alpha(pot_x_range=self.x_range, dx=self.dx)
             self._potential += alpha_array * (
-                state.kT * np.log(current_dist[:, 1] / target_dist[:, 1]) / N
+                state.kT * np.log(current_dist[:, 1] / target_dist[:, 1])
             )
+            state_count += 1
+        self._potential /= state_count
         # Apply corrections to regions without distribution overlap
         if isinstance(self, msibi.forces.Pair):
             self._potential, head_cut, tail_cut, real_indices = pair_corrections(
