### 2026-03-09 cd54b9c
Add tests for ignore states
