+++ msibi/tests/test_forces.py
@@ -96,6 +96,14 @@ class TestForce(BaseTest):
         assert bond2.nbins == bond.nbins == 60
         os.remove("test.csv")
 
+    def test_ignore_states(self, msibi, stateX, stateY):
+        pair = Pair(type1="A", type2="B", optimize=True, nbins=60)
+        assert pair.ignore_states == []
+
+        pair = Pair(type1="A", type2="B", optimize=True, nbins=60, ignore_states=[stateY])
+        assert len(pair.ignore_states) == 1
+        assert pair.ignore_states[0] == stateY
+
     def test_fit_scores(self, msibi, stateX, stateY):
         msibi.gsd_period = 10
         bond = Bond(
