+++ msibi/forces.py
@@ -89,6 +89,12 @@ class Force:
         to both the head and tail of the potential.
         Non-bonded forces (``Pair``) only apply this correction to the head of
         the potential.
+    ignore_states : List, optional
+        This lets you exclude certain state points to be used in updating
+        this potential. This may be helpful where a state point has a 
+        configuration such that the RDF of a pair is zero at all values. 
+        For example, a bilayer system where the head and tail beads 
+        are outside the cutoff distance for the entire trajectory.
     """
 
     def __init__(
@@ -784,6 +790,12 @@ class Bond(Force):
         The type of correciton form to apply to the potential.
         This is only used when the Force is set to be optimized.
         This correction is applied to both the head and tail of the potential.
+    ignore_states : List, optional
+        This lets you exclude certain state points to be used in updating
+        this potential. This may be helpful where a state point has a 
+        configuration such that the RDF of a pair is zero at all values. 
+        For example, a bilayer system where the head and tail beads 
+        are outside the cutoff distance for the entire trajectory.
     """
 
     def __init__(
@@ -797,6 +809,7 @@ class Bond(Force):
         correction_fit_window: Optional[int] = 10,
         maxfev: Optional[int] = 1000,
         correction_form: Callable = harmonic,
+        ignore_states: List = None 
     ):
         self.type1, self.type2 = sorted([type1, type2], key=natural_sort)
         name = f"{self.type1}-{self.type2}"
@@ -809,6 +822,7 @@ class Bond(Force):
             correction_fit_window=correction_fit_window,
             maxfev=maxfev,
             correction_form=correction_form,
+            ignore_states=ignore_states
         )
 
     def set_harmonic(self, r0: Union[float, int], k: Union[float, int]) -> None:
@@ -925,6 +939,12 @@ class Angle(Force):
         The type of correciton form to apply to the potential.
         This is only used when the Force is set to be optimized.
         This correction is applied to both the head and tail of the potential.
+    ignore_states : List, optional
+        This lets you exclude certain state points to be used in updating
+        this potential. This may be helpful where a state point has a 
+        configuration such that the RDF of a pair is zero at all values. 
+        For example, a bilayer system where the head and tail beads 
+        are outside the cutoff distance for the entire trajectory.
     """
 
     def __init__(
@@ -939,6 +959,7 @@ class Angle(Force):
         correction_fit_window: Optional[int] = 10,
         maxfev: Optional[int] = 1000,
         correction_form: Callable = harmonic,
+        ignore_states: List = None 
     ):
         self.type1 = type1
         self.type2 = type2
@@ -953,6 +974,7 @@ class Angle(Force):
             correction_fit_window=correction_fit_window,
             maxfev=maxfev,
             correction_form=correction_form,
+            ignore_states=ignore_states
         )
 
     def set_harmonic(self, t0: Union[float, int], k: Union[float, int]) -> None:
@@ -1075,6 +1097,12 @@ class Pair(Force):
         The type of correciton form to apply to the potential.
         This is only used when the Force is set to be optimized.
         This correction is only applied to both the head of the potential.
+    ignore_states : List, optional
+        This lets you exclude certain state points to be used in updating
+        this potential. This may be helpful where a state point has a 
+        configuration such that the RDF of a pair is zero at all values. 
+        For example, a bilayer system where the head and tail beads 
+        are outside the cutoff distance for the entire trajectory.
     """
 
     def __init__(
@@ -1091,6 +1119,7 @@ class Pair(Force):
         maxfev: Optional[int] = 1000,
         exclude_bonded: bool = False,
         head_correction_form: Callable = exponential,
+        ignore_states: List = None 
[…]
