### 2026-03-12 8ec72ab
include tps in sim printout
