### 2026-03-12 bb032a8
add method to create state-specific params for Forces
