+++ msibi/forces.py
@@ -179,9 +179,10 @@ class Force:
             raise PotentialNotOptimizedError("apply a smoothing window")
         if not isinstance(value, int) or value <= 0:
             raise ValueError("The smoothing window must be an integer > 0.")
+        # Update the stored target distribution
         self._smoothing_window = value
         for state in self._states:
-            self._add_state(state)
+            self._update_target_distribution(state)
 
     @property
     def smoothing_order(self) -> int:
@@ -195,9 +196,10 @@ class Force:
             raise PotentialNotOptimizedError("apply a smoothing order")
         if not isinstance(value, int) or value <= 0:
             raise ValueError("The smoothing order must be an integer.")
+        # Update the stored target distribution
         self._smoothing_order = value
         for state in self._states:
-            self._add_state(state)
+            self._update_target_distribution(state)
 
     @property
     def nbins(self) -> int:
@@ -211,7 +213,7 @@ class Force:
             raise ValueError("nbins must be an integer.")
         self._nbins = value
         for state in self._states:
-            self._add_state(state)
+            self._update_target_distribution(state)
 
     def smooth_potential(self) -> None:
         """Smooth and overwrite the current potential.
@@ -482,7 +484,7 @@ class Force:
     def set_target_distribution(
         self, state: msibi.state.State, array: np.ndarray
     ) -> None:
-        """Store the target distribution for a given state.
+        """Manually set the target distribution for a given state.
 
         Parameters
         ----------
@@ -616,7 +618,7 @@ class Force:
         state : msibi.state.State
             Instance of a State object previously created.
         """
-        if self.optimize:
+        if self.optimize and state not in self.ignore_states:
             target_distribution = self._get_state_distribution(state=state, query=False)
             if self.smoothing_window and self.smoothing_order:
                 target_distribution[:, 1] = savgol_filter(
@@ -636,6 +638,27 @@ class Force:
             "distribution_history": [],
             "path": state.dir,
         }
+        self._states[state].update(self._state_defaults())
+        # Apply user-set params overriding defaults
+        if state in self._pending_state_params:
+            self._states[state].update(self._pending_state_params[state])
+
+    def _state_defaults(self):
+        """Needed place holder, this method is only used in Pair."""
+        return {}
+
+    def _update_target_distribution(self, state: msibi.state.State) -> None:
+            target_distribution = self._get_state_distribution(state=state, query=False)
+            if self.smoothing_window and self.smoothing_order:
+                target_distribution[:, 1] = savgol_filter(
+                    x=target_distribution[:, 1],
+                    window_length=self.smoothing_window,
+                    polyorder=self.smoothing_order,
+                )
+                neg_indices = np.where(target_distribution[:, 1] < 0)[0]
+                target_distribution[:, 1][neg_indices] = 0
+            self._states[state]["target_distribution"] = target_distribution
+
 
     def _compute_current_distribution(self, state: msibi.state.State) -> None:
         """Find the current distribution of the query trajectory.
@@ -867,6 +890,10 @@ class Bond(Force):
         }
         return table_entry
 
+    def _state_defaults(self):
+        """Needed for Force._add_state()."""
+        return {"optimize_against": self.optimize}
+
     def _get_distribution(self, state: msibi.state.State, gsd_file: str) -> np.ndarray:
         """Calculate a bond length distribution.
 
@@ -1014,6 +1041,10 @@ class Angle(Force):
         table_entry = {"U": self.potential, "tau": self.force}
         return table_entry
 
[…]
