+++ msibi/tests/base_test.py
@@ -56,7 +56,7 @@ class BaseTest:
             alpha0=1.0,
             kT=4.0,
             traj_file=os.path.join(test_assets, "AB-4.0kT.gsd"),
-            n_frames=100,
+            n_frames=20,
             sampling_stride=1,
             _dir=tmp_path,
         )
