+++ msibi/tests/test_forces.py
@@ -340,6 +340,7 @@ class TestPair(BaseTest):
             optimize=True,
             exclude_bond_depth=2,
         )
+        pair.set_lj(sigma=1, epsilon=1, r_cut=3, r_min=0.1)
         pair._add_state(stateX)
         assert pair._states[stateX]["exclude_bond_depth"] == 2
         assert pair._states[stateX]["optimize_against"] == True 
@@ -353,7 +354,8 @@ class TestPair(BaseTest):
             optimize=True,
             exclude_bond_depth=2,
         )
-        pair.set_state_params(stateX, exclude_bond_depth=3, optimize=False)
+        pair.set_lj(sigma=1, epsilon=1, r_cut=3, r_min=0.1)
+        pair.set_state_params(stateX, exclude_bond_depth=3, optimize_against=False)
         pair._add_state(stateX)
         assert pair._states[stateX]["exclude_bond_depth"] == 3 
         assert pair._states[stateX]["optimize_against"] == False 
