+++ environment-dev.yml
@@ -2,7 +2,7 @@ name: msibi-dev
 channels:
 - conda-forge
 dependencies:
-- cmeutils >=1.4
+- cmeutils =1.4
 - freud >=3.4
 - gsd >=3.0
 - hoomd >=5.4
