### 2026-03-19 3bd5f8f
minor changes in paper.md
