### 2026-03-19 #9244 @danielskatz
👋 @chrisjonesBSU - I've suggested some minor changes after proofreading - see https://github.com/mosdef-hub/msibi/pull/72. Please merge this, or let me know what you disagree with.
