### 2026-03-19 #9244 @chrisjonesBSU
> 👋 [@chrisjonesBSU](https://github.com/chrisjonesBSU) - I've suggested some minor changes after proofreading - see [mosdef-hub/msibi#72](https://github.com/mosdef-hub/msibi/pull/72). Please merge this, or let me know what you disagree with.

Thank you! It has been merged.
