### 2026-03-19 #9244 @chrisjonesBSU
@danielskatz

I used a tagged release with version 0.3.1

https://zenodo.org/records/19111404 and the DOI is 10.5281/zenodo.19111404
