### 2026-03-19 #9244 @danielskatz
Congratulations to @chrisjonesBSU (Chris D. Jones) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers - we couldn't do this without you
